feat: use remote rpc instead of terminal cmd
